pyproject.toml
[tool.poetry]
name = "wavemap_paper"
version = "0.1.11"
description = ""
authors = ["Eric Kenji Lee <erickenjilee@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
umap-learn = "^0.5.3"
jupyterlab = "^3.5.1"
ipykernel = "^6.19.2"
shap = "^0.41.0"
scikit-learn = "^1.2.0"
xgboost = "^1.7.2"
cylouvain = "^0.2.2"
matplotlib = "^3.6.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
